# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Dochub API
  version: "1.0.0"
  description: |
    Organizational RAG document hub.

    **Nothing reaches the backend while a user is choosing sources.** The browser holds
    repository details, folder ids and local files until Process is pressed.

    **Process contract.** `POST /artifacts/{artifactId}/process` is the only call the
    Upload screen makes. For each source it writes a `source_documents` row with status
    `RequestUpload` and puts the source on the extract queue
    (`dochub-source-upload-requested`). The extractor service claims one message at a
    time, flips the row to `Uploading`, moves the files to Azure Blob storage under
    `teams/{team}/groups/{group}/artifacts/{artifact}/{uploadedAt}`, sets `Uploaded`, and
    hands the documents to the process queue (`dochub-document-process-requested`). For
    every blob it stores the content MD5 Azure returned in a `document_versions` row.
    The RAG platform (`ragplatform/`) consumes the process queue, indexes each document
    and reports back through the **Ingestion** endpoints: once per document, then once
    for the artifact, which closes the source and sets the artifact's status.

    Local files are the one exception that needs a prior call: their bytes go to
    `POST /uploads/staging` immediately before Process, and the returned staging ids are
    passed in the source's `options`.

    **Recurring updates.** A SharePoint or Google Drive source may carry a `schedule`,
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/notifications:
    get:
      tags: [Notifications]
      summary: Notifications for the caller, newest first
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
        - { name: unreadOnly, in: query, schema: { type: boolean } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Notification" } }
                  unreadCount: { type: integer }
  /api/notifications/read:
    post:
      tags: [Notifications]
      summary: Mark notifications read
      description: Send `null` or an empty body to mark everything read.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: array
              nullable: true
              items: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, properties: { markedRead: { type: integer } } }
...
  /api/ingestion/documents/{documentId}/processed:
    post:
      tags: [Ingestion]
      summary: Report one document indexed
      security: [{ serviceKey: [] }]
      parameters:
        - { name: documentId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DocumentProcessedRequest" }
      responses:
        "200":
          description: Document is now Indexed.
          content: { application/json: { schema: { $ref: "#/components/schemas/IngestionAck" } } }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409":
          description: |
            `stale_version` — a newer upload of this file has replaced the version being
            reported; the caller should treat it as done. `source_mismatch` — the document
            is not part of that source.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "503": { description: "`Ingestion:ServiceKey` is not configured." }
  /api/ingestion/documents/{documentId}/failed:
    post:
      tags: [Ingestion]
      summary: Report one document failed
      security: [{ serviceKey: [] }]
      parameters:
        - { name: documentId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DocumentFailedRequest" }
      responses:
        "200":
          description: Failure recorded; the error is shown on the document.
          content: { application/json: { schema: { $ref: "#/components/schemas/IngestionAck" } } }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "503": { description: "`Ingestion:ServiceKey` is not configured." }
  /api/ingestion/artifacts/{artifactId}/processed:
    post:
      tags: [Ingestion]
      summary: Report a source batch finished
      description: |
        Closes the source: documents never reported are marked Failed, the source becomes
        `Processed`, `PartiallyFailed` or `Failed`, and the artifact status is computed from
        **all** of the artifact's documents. Raises the completion notification.
      security: [{ serviceKey: [] }]
      parameters:
        - { name: artifactId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ArtifactProcessedRequest" }
      responses:
        "200":
          description: Batch closed.
          content: { application/json: { schema: { $ref: "#/components/schemas/ArtifactProcessedResponse" } } }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "503": { description: "`Ingestion:ServiceKey` is not configured." }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token from `POST /api/auth/sso`. Carries the active `org_id`, so every request is tenant-scoped by the token.
    serviceKey:
      type: apiKey
      in: header
      name: X-Dochub-Service-Key
      description: Equal to `Ingestion:ServiceKey` (≥32 chars). Used only by the RAG platform.
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
  schemas:
    ApiError:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: owner_only }
        message: { type: string }
        details: { nullable: true }
...
    SourceDocumentStatus:
      type: string
      description: |
        The whole journey of one submitted source.
        RequestUpload → Uploading → Uploaded → Processing → Processed
      enum: [RequestUpload, Uploading, Uploaded, Processing, Processed, PartiallyFailed, Failed, Cancelled]
...
    ArtifactStatus:
      type: string
      description: Stored on the artifact. Set to Pending on Process, Processing once files reach the process queue, and finally by the artifact callback.
      enum: [Empty, Pending, Processing, Processed, PartiallyProcessed, Failed]
...
    DocumentProcessedRequest:
      type: object
      required: [sourceDocumentId, chunkCount]
      properties:
        sourceDocumentId: { type: string, format: uuid }
        documentVersionId:
          type: string
          format: uuid
          nullable: true
          description: The version that was indexed. A mismatch with the current version returns 409 stale_version.
        chunkCount: { type: integer }
        contentSha256: { type: string, nullable: true }
        embeddingModel: { type: string, nullable: true }
        pipelineVersion: { type: string, nullable: true }
        skipped: { type: boolean, description: True when the content was already indexed and no work was done. }
    DocumentFailedRequest:
      type: object
      required: [sourceDocumentId, error]
      properties:
        sourceDocumentId: { type: string, format: uuid }
        documentVersionId: { type: string, format: uuid, nullable: true }
        error: { type: string, example: "detect: unsupported_in_this_build: pdf extraction arrives in phase 3" }
        stage: { type: string, nullable: true, example: detect }
        permanent: { type: boolean }
    ArtifactProcessedRequest:
      type: object
      required: [sourceDocumentId]
      properties:
        sourceDocumentId: { type: string, format: uuid, description: The source batch being closed. }
    IngestionAck:
      type: object
      properties:
        status: { type: string }
        message: { type: string }
    ArtifactProcessedResponse:
      type: object
      properties:
        artifactStatus: { $ref: "#/components/schemas/ArtifactStatus" }
        sourceStatus: { $ref: "#/components/schemas/SourceDocumentStatus" }
        processed: { type: integer }
        failed: { type: integer }
        total: { type: integer }
...
    Notification:
      type: object
      properties:
        id: { type: string, format: uuid }
        type: { type: string, example: processing.completed }
        title: { type: string }
        body: { type: string }
        severity: { type: string, enum: [Info, Success, Warning, Error] }
        entityType: { type: string, nullable: true }
        entityId: { type: string, format: uuid, nullable: true }
        read: { type: boolean }
        createdAt: { type: string, format: date-time }